Social speed reader and database creator.
The purpose Spudy is to help students(or anyone who needs to read
extensively) learn their information quickly while retaining high levels
of comprehension. It utilizes several modern api's to convert images of
textbook/book pages to plain text which is then displayed to the user at
an adjustable reading rate of up to 800 words per minute. By displaying
one word at a time, the user doesnt have to move his or her eyes and
interprets the words as pictures rather than letter-by-letter, allowing
for much faster reading. We strongly believe that such a tool is
absolutely essential for students who have limited time to cover very
large amounts of material on a day-to-day basis. The expansion potential
for such an app is unbounded.